figured out a critical roadblock.
my biggest problem was in order to be like Plex, the location of your manga library needs to be configurable. but I wasn't sure how to serve files from a location outside of static. many people would suggest nginx, but Plex doesn't ask you to install that, so I don't want to either. a lot needs to be improved and implemented, but the road ahead is much clearer to me.